Moving token to module
