+++ backend/auth-service/pom.xml
@@ -15,12 +15,19 @@
     <description>Auth service</description>
 
     <properties>
+
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>com.ifood</groupId>
             <artifactId>core</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.ifood</groupId>
+            <artifactId>token</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
