+++ backend/auth-service/src/main/resources/application.yml
@@ -15,5 +15,4 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:8081/eureka/
-    fetch-registry: true
     register-with-eureka: true
