+++ backend/customer-service/src/main/resources/application.yml
@@ -15,7 +15,6 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:8081/eureka/
-    fetch-registry: true
     register-with-eureka: true
