+++ backend/pom.xml
@@ -28,6 +28,7 @@
         <module>customer-service</module>
         <module>core</module>
         <module>auth-service</module>
+        <module>token</module>
     </modules>
 
     <dependencies>
