+++ backend/token/src/main/java/com/ifood/token/security/config/SecurityTokenConfig.java
@@ -1,0 +1,36 @@
+package com.ifood.token.security.config;
+
+import com.ifood.core.property.JwtConfiguration;
+import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.web.cors.CorsConfiguration;
+
+import javax.servlet.http.HttpServletResponse;
+
+import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
+
+@AllArgsConstructor
+public class SecurityTokenConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    protected JwtConfiguration jwtConfiguration;
+
+    @Override
+    protected void configure (HttpSecurity http) throws Exception {
+        http
+                .csrf().disable()
+                .cors().configurationSource(request -> new CorsConfiguration().applyPermitDefaultValues())
+                .and()
+                .sessionManagement().sessionCreationPolicy(STATELESS)
+                .and()
+                .exceptionHandling().authenticationEntryPoint((req, resp, e) -> resp.sendError(HttpServletResponse.SC_UNAUTHORIZED))
+                .and()
+                .authorizeRequests()
+                .antMatchers(jwtConfiguration.getLoginUrl()).permitAll()
+                .antMatchers("/admin/**").hasRole("ADMIN")
+                .anyRequest().authenticated();
+    }
+
+}
